pleroma-fox-tan-shy.png: Add another version of pleroma-tan
Commissioned by https://shitposter.club/users/moonman and
drawn by https://shpposter.club/users/Nightingalle

Intended to be used when mastofe crashes out and allows to report a bug
among other things.

Should require a bit of modification on mastofe to use the image as it's more
of a vanilla flavour thing.