Remove No MX lookups setting
This setting defaults to false so the relay host will be used in an MX query so
multiple SMTP servers can be used. gen_smtp code states that all records returned from the
MX query are attempted in order and only a permanent SMTP error will stop the client from
attempting other servers in the list. Connection failures, TLS issues, etc will cause it to
try the next host.

If there is no MX record associated with the relay host, it automatically tries connecting to it
directly.

There is really no reason to expose this to end users. The default value is fine for everyone.